move timings to state to break dep cycle
