+++ pkg/feature/feature.go
@@ -31,8 +31,6 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 	"knative.dev/pkg/apis"
 	"knative.dev/pkg/injection/clients/dynamicclient"
-	"knative.dev/reconciler-test/pkg/environment"
-
 	"knative.dev/reconciler-test/pkg/state"
 )
 
@@ -229,7 +227,7 @@ func DeleteResources(ctx context.Context, t T, refs []corev1.ObjectReference) er
 
 	var lastResource corev1.ObjectReference // One still present resource
 
-	interval, timeout := environment.PollTimingsFromContext(ctx)
+	interval, timeout := state.PollTimingsFromContext(ctx)
 	err := wait.Poll(interval, timeout, func() (bool, error) {
 		for _, ref := range refs {
 			gv, err := schema.ParseGroupVersion(ref.APIVersion)
